Filtering params update
Represents the amount of filtering that can be done to streams on Twitter's side. According to Twitter's documentation, "When displaying a stream of Tweets to end users (dashboards or live feeds at a presentation or conference, for example) it is suggested that you set this value to medium. Custom is convenient to further enhance filtering abilities and adapt NLU properties here in the future