Guarantee bartering with memory location
Attempt to resolve the next item in the stresam. It will yield Some(Val) every time a value is sent from the sender's end.Note that because next doesn't take ownership over the stream, the Stream type must be Unpin. If you want to use next with a !Unpin stream, you'll first have to pin the stream.